+++ R/get_mem_lim_m.R
@@ -1,14 +1,15 @@
 # an internal constant shared by this function and get_mem_lim
 GB <- 1024*1024*1024
 
-get_mem_lim_m <- function(m = NA, n = NA, mem = NA, factor = 0.7, verbose = FALSE) {
+# mem is in GB
+get_mem_lim_m <- function(n, m = NA, mem = NA, factor = 0.7, verbose = FALSE) {
     # NOTE: in estimating the chunk size, we don't know ahead of time if there are NAs or not!
     # so these calculations assume M in getMAInt is an n*n matrix
     # if there aren't any NAs, M is a scalar and we end up underestimating memory usage (better than the other way around)
 
     # n must be defined now, or this doesn't work!
-    if (is.na(n))
-        stop('n must be defined!')
+    if (missing(n))
+        stop('`n` is required!')
     
     # try to get total memory from the system if mem wasn't specified, so it works reasonably
     if (is.na(mem)) {
@@ -36,25 +37,18 @@ get_mem_lim_m <- function(m = NA, n = NA, mem = NA, factor = 0.7, verbose = FALS
     # = 8*(m*n + n*n + 20)
     # given fixed n, solve for m:
     # get maximum m (number of SNPs) given n and the memory requested
-    mc <- (mem/8 - 20 - n*n)/n
+    data <- solve_m_mem_lim(
+        mem = mem,
+        n = n,
+        m = m,
+        mat_m_n = 1, # X (0.5) + ?
+        mat_n_n = 1 # A + M (0.5 + 0.5)
+    )
+    #    m_chunk <- (mem/8 - 20 - n*n)/n
 
-    # NOTE m may be missing if X is a function, so we can't make these simplifying decisions (to balance load) without m in that case...
-    if (!is.na(m)) {
-        if (m < mc) {
-            mc <- m # use the smaller one
-        } else {
-            # should "redistribute" based on number of chunks, to lower memory even more per iteration
-            mc <- ceiling( m/ceiling(m/mc) ) # this lowers mc even more, balances load better
-        }
-    }
+    if (verbose)
+        message('Choice of chunk size should limit mem to about ', round( data$mem_chunk / GB, 2 ), ' GB')
 
-    if (verbose) {
-        # actual memory in use, in bytes
-        mem_act <- 8*(mc*n + n*n + 20)
-        
-        message('Choice of chunk size should limit mem to about ', round( mem_act / GB, 2 ), ' GB')
-    }
-
-    return(mc) # return desired value
+    return(data$m_chunk) # return desired value
 }